requirements.txt
pytest >= 7.1.2
z3-solver >= 4.12.1
toolz >= 0.12.0
networkx >= 2.8.4
antlr4-python3-runtime == 4.12.0
pygraphviz >= 1.9
graphviz >= 0.20.1
pandas >= 2.1.1
loguru >= 0.7.2
numpy == 2.0.1
matplotlib >= 3.10.0
line_profiler_pycharm >= 1.1.0

# pytest
pytest-assume >= 2.4.3
pytest-xdist >= 3.5.0
pytest-timeout >= 2.2.0

# egglog
egglog == 6.0.0
anywidget >= 0.9.2

# other
paramiko >= 2.9.3
scp >=0.15.0

botorch == 0.15.1
autograd == 1.8.0
gpytorch == 1.14
scikit-learn == 1.7.2
scipy == 1.16.0
seaborn == 0.13.2
torch == 2.8.0
tqdm == 4.67.1